refactor: 重构重复资源判断逻辑，添加缺失的包需求

+++ collect_data.py
@@ -11,6 +11,9 @@ version_hash = utils.try_get_version_hash()
 
 logger = logger_helper.get_logger(logging.INFO)
 session = requests.Session()
+# session.proxies.update({
+#     'https': 'http://127.0.0.1:12450'
+# })
 
 METADATA_TEMPLATE = utils.get_b64_data(b'aHR0cHM6Ly95b3N0YXItc2VydmVyaW5mby5ibHVlYXJjaGl2ZXlvc3Rhci5jb20ve30uanNvbg==')
 A_BUNDLE_CATALOG_TEMPLATE = utils.get_b64_data(b'e30vQW5kcm9pZC9idW5kbGVEb3dubG9hZEluZm8uanNvbg==')
@@ -34,28 +37,29 @@ for d in dirs:
 
 def collect_bundle_files(base_url: str, full_json: dict):
     con = sqlite3.connect(utils.DATABASE_NAME)
-    dl_dict = full_json[utils.get_b64_data(b'QnVuZGxlRmlsZXM=')]
+    dl_list = full_json[utils.get_b64_data(b'QnVuZGxlRmlsZXM=')]
     bundle_base_url = A_BUNDLE_BASE_TEMPLATE.format(base_url)
-    success_cnt, skip_cnt, total_cnt, progress_cnt = 0, 0, len(dl_dict), 0
+    success_cnt, skip_cnt, total_cnt, progress_cnt = 0, 0, len(dl_list), 0
     cur = con.cursor()
-    for bundle_data in dl_dict:
+    all_database_dict = {bundle_crc: bundle_full_path for bundle_crc, bundle_full_path
+                         in cur.execute("SELECT CRC, FILE_FULL_PATH FROM bundle_dict").fetchall()}
+    for bundle_data in dl_list:
         progress_cnt += 1
-        table_row_data = cur.execute(
-            "SELECT * FROM bundle_dict WHERE FILE_NAME=(?) AND CRC=(?)",
-            (bundle_data['Name'], bundle_data['Crc'])
-        ).fetchone()
-        if not table_row_data:
-            save_path = os.path.join(BUNDLE_FILES_DIR, bundle_data['Name'])
-            logger.debug(f'collecting bundle file: {bundle_data["Name"]} ...')
+        local_file_path = all_database_dict.get(bundle_data['Crc'])
+        if not local_file_path or not os.path.exists(local_file_path):
+            save_path = os.path.join(BUNDLE_FILES_DIR, bundle_data['Name']) if not local_file_path else local_file_path
+            logger.debug(f'{"collecting" if not local_file_path else "repairing"} bundle file: {bundle_data["Name"]} ...')
+            logger.debug(f'from url: {bundle_base_url}{bundle_data["Name"]}')
             data = session.get(f'{bundle_base_url}{bundle_data["Name"]}').content
             with open(save_path, 'wb') as f:
                 f.write(data)
             cur.execute('''
                 INSERT INTO bundle_dict (FILE_ID, FILE_FULL_PATH, SIZE, CRC, FILE_NAME, UPDATE_TIME, VERSION_MARK)
                  VALUES (?, ?, ?, ?, ?, ?, ?)
-            ''', (None, save_path, bundle_data['Size'], bundle_data['Crc'], bundle_data['Name'], utils.get_cur_time(), version_hash))
+            ''', (None, save_path, bundle_data['Size'], bundle_data['Crc'], bundle_data['Name'], utils.get_cur_time(),
+                  version_hash))
             con.commit()
-            logger.info(f'({success_cnt+skip_cnt}/{total_cnt}) bundle file: {bundle_data["Name"]} collected.')
+            logger.info(f'({progress_cnt}/{total_cnt}) bundle file: {bundle_data["Name"]} {"collected" if not local_file_path else "repaired"}.')
             success_cnt += 1
         else:
             logger.debug(f'bundle file: {bundle_data["Name"]} skipped.')
@@ -66,44 +70,41 @@ def collect_bundle_files(base_url: str, full_json: dict):
     con.close()
     return success_cnt, skip_cnt, total_cnt
 
-def cmp_table_row_data(all_data):
-    def no_current_binary_key(item):
-        binary_data = item[1]
-        for line in all_data:
-            crc = line[4]
-            filename = line[7]
-            if crc == binary_data['Crc'] and filename == binary_data['fileName']:
-                return False
-        return True
-    return no_current_binary_key
 
 def collect_binary_files(base_url: str, full_json: dict):
     con = sqlite3.connect(utils.DATABASE_NAME)
     dl_dict = full_json[utils.get_b64_data(b'VGFibGU=')]
     binary_base_url = BINARY_BASE_TEMPLATE.format(base_url)
     cur = con.cursor()
-    all_data = cur.execute("SELECT * FROM binary_dict").fetchall()
-    rest_binary_data = list(filter(cmp_table_row_data(all_data), dl_dict.items()))
-    success_cnt, skip_cnt, total_cnt, progress_cnt = 0, len(dl_dict) - len(rest_binary_data), len(rest_binary_data), 0
+    all_database_dict = {binary_crc: binary_full_path for binary_crc, binary_full_path
+                         in cur.execute("SELECT CRC, FILE_FULL_PATH FROM binary_dict").fetchall()}
+    success_cnt, skip_cnt, total_cnt, progress_cnt = 0, 0, len(dl_dict), 0
 
-    for item in rest_binary_data:
+    for binary_key, binary_data in dl_dict.items():
         progress_cnt += 1
-        binary_data = item[1]
-        save_path = f'{BINARY_FILES_DIR}/{binary_data["path"]}'
-        save_dir = os.path.dirname(save_path)
-        if not os.path.exists(save_dir):
-            os.makedirs(os.path.dirname(save_path), exist_ok=True)
-        logger.debug(f'collecting binary file: {binary_data["fileName"]} ...')
-        data = session.get(f'{binary_base_url}{binary_data["path"]}').content
-        with open(save_path, 'wb') as f:
-            f.write(data)
-        cur.execute('''
-            INSERT INTO binary_dict (FILE_ID, FILE_FULL_PATH, SOURCE_FILE_PATH, SIZE, CRC, BYTES, MIDIA_TYPE, FILE_NAME, UPDATE_TIME, VERSION_MARK)
-                VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-        ''', (None, save_path, binary_data['path'], binary_data['bytes'], binary_data['Crc'], binary_data['bytes'], binary_data['mediaType'], binary_data['fileName'], utils.get_cur_time(), version_hash))
-        con.commit()
-        logger.info(f'({success_cnt+skip_cnt}/{total_cnt}) binary file: {binary_data["fileName"]} collected.')
-        success_cnt += 1
+        local_file_path = all_database_dict.get(binary_data['Crc'])
+        if not local_file_path or not os.path.exists(local_file_path):
+            save_path = f'{BINARY_FILES_DIR}/{binary_data["path"]}' if not local_file_path else local_file_path
+            save_dir = os.path.dirname(save_path)
+            if not os.path.exists(save_dir):
+                os.makedirs(os.path.dirname(save_path), exist_ok=True)
+            logger.debug(f'{"collecting" if not local_file_path else "repairing"} binary file: {binary_data["fileName"]} ...')
+            logger.debug(f'from url: {binary_base_url}{binary_data["path"]}')
+            data = session.get(f'{binary_base_url}{binary_data["path"]}').content
+            with open(save_path, 'wb') as f:
+                f.write(data)
+            cur.execute('''
+                INSERT INTO binary_dict (FILE_ID, FILE_FULL_PATH, SOURCE_FILE_PATH, SIZE, CRC, BYTES, MIDIA_TYPE, FILE_NAME, UPDATE_TIME, VERSION_MARK)
+                 VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+            ''', (None, save_path, binary_data['path'], binary_data['bytes'], binary_data['Crc'], binary_data['bytes'],
+                  binary_data['mediaType'], binary_data['fileName'], utils.get_cur_time(), version_hash))
+            con.commit()
+            logger.info(
+                f'({progress_cnt}/{total_cnt}) binary file: {binary_data["fileName"]} {"collected" if not local_file_path else "repaired"}.')
+            success_cnt += 1
+        else:
+            logger.debug(f'binary file: {binary_data["fileName"]} skipped.')
+            skip_cnt += 1
 
     logger.info(f'Task of collecting binary files has finished.')
     logger.info(f'success: {success_cnt}, skipped: {skip_cnt}')
@@ -118,9 +119,8 @@ def collect_table_files(base_url: str, full_json: dict):
     table_base_url = TABLE_BASE_TEMPLATE.format(base_url)
     success_cnt, skip_cnt, total_cnt, progress_cnt = 0, 0, len(dl_dict), 0
     cur = con.cursor()
-    for table_name in dl_dict:
+    for table_name, table_data in dl_dict.items():
         progress_cnt += 1
-        table_data = dl_dict[table_name]
         save_path = os.path.join(TABLE_FILES_DIR, table_name)
         table_row_data = cur.execute(
             "SELECT * FROM table_dict WHERE FILE_NAME=(?) AND CRC=(?)",
@@ -134,9 +134,10 @@ def collect_table_files(base_url: str, full_json: dict):
             cur.execute('''
                 INSERT INTO table_dict (FILE_ID, FILE_FULL_PATH, SIZE, CRC, FILE_NAME, UPDATE_TIME, VERSION_MARK)
                  VALUES (?, ?, ?, ?, ?, ?, ?)
-            ''', (None, save_path, table_data['Size'], table_data['Crc'], table_name, utils.get_cur_time(), version_hash))
+            ''', (
+                None, save_path, table_data['Size'], table_data['Crc'], table_name, utils.get_cur_time(), version_hash))
             con.commit()
-            logger.info(f'({success_cnt+skip_cnt}/{total_cnt}) table file: {table_name} collected.')
+            logger.info(f'({progress_cnt}/{total_cnt}) table file: {table_name} collected.')
             success_cnt += 1
         else:
             logger.debug(f'table file: {table_name} skipped.')
@@ -148,7 +149,8 @@ def collect_table_files(base_url: str, full_json: dict):
     return success_cnt, skip_cnt, total_cnt
 
 
-def check_and_create_catalog_file(catalog_type: str, catalog_file_name: str, catalog_json: dict, con: sqlite3.Connection):
+def check_and_create_catalog_file(catalog_type: str, catalog_file_name: str, catalog_json: dict,
+                                  con: sqlite3.Connection):
     json_md5 = hashlib.md5(json.dumps(catalog_json).encode()).hexdigest()
     logger.debug(f'{catalog_type} type catalog\'s md5 is: {json_md5}')
     catalog_full_path = con.execute(
@@ -181,21 +183,25 @@ def collect_all_catalogs(use_local_file=False):
     con = sqlite3.connect(utils.DATABASE_NAME)
     cur = con.cursor()
     # collect main catalog
-    main_catalog_json, file_source = utils.get_newest_catalog('MAIN', cur, METADATA_TEMPLATE.format(version_hash), session, use_local_file)
+    main_catalog_json, file_source = utils.get_newest_catalog('MAIN', cur, METADATA_TEMPLATE.format(version_hash),
+                                                              session, use_local_file)
     if file_source == 'remote':
         check_and_create_catalog_file('MAIN', 'main-catalog', main_catalog_json, con)
     base_url = utils.get_base_from_json(main_catalog_json)
 
     # collect child catalogs
-    bundle_catalog_json, file_source = utils.get_newest_catalog('BUN', cur, A_BUNDLE_CATALOG_TEMPLATE.format(base_url), session, use_local_file)
+    bundle_catalog_json, file_source = utils.get_newest_catalog('BUN', cur, A_BUNDLE_CATALOG_TEMPLATE.format(base_url),
+                                                                session, use_local_file)
     if file_source == 'remote':
         check_and_create_catalog_file('BUN', 'bundle-catalog', bundle_catalog_json, con)
 
-    binary_catalog_json, file_source = utils.get_newest_catalog('BIN', cur, BINARY_CATALOG_TEMPLATE.format(base_url), session, use_local_file)
+    binary_catalog_json, file_source = utils.get_newest_catalog('BIN', cur, BINARY_CATALOG_TEMPLATE.format(base_url),
+                                                                session, use_local_file)
     if file_source == 'remote':
         check_and_create_catalog_file('BIN', 'binary-catalog', binary_catalog_json, con)
 
-    table_catalog_json, file_source = utils.get_newest_catalog('TABLE', cur, TABLE_CATALOG_TEMPLATE.format(base_url), session, use_local_file)
+    table_catalog_json, file_source = utils.get_newest_catalog('TABLE', cur, TABLE_CATALOG_TEMPLATE.format(base_url),
+                                                               session, use_local_file)
     if file_source == 'remote':
         check_and_create_catalog_file('TABLE', 'table-catalog', table_catalog_json, con)
